adding basic token verificatino check for each query instead of createUser for demonstration purposes

+++ app/graphql/resolvers/delete_product.rb
@@ -3,6 +3,7 @@ class Resolvers::DeleteProduct < GraphQL::Function
 
 	type Types::ProductType
 	def call(_obj, args, _ctx)
+		return GraphQL::ExecutionError.new("Not Authorized to make this request") unless ctx[:session][:token]
 		return GraphQL::ExecutionError.new("Not Authorized to make this request") unless ctx[:current_user].role == "[\"ADMIN_USER\"]"
 		product = Product.find(args[:id])
 		product.destroy
